Tested pdf conversion with wkhtmltopdf.exe
This works!  I tried really hard to get the c library bindings to
work, but I think I need to compile from source (which requires
a decent amount of setup that I don't want to do right now).  For
now, having the wkhtmltopdf executable in path or in the same
directory as the output .exe for this project will work.