BUG ENCOUNTERED: Added code to add .png files
Currently it appears that adding the <img> tag is working correctly,
however my tests fail because the generated webpage won't load the
image and instead loads the alt text (despite the image having
a valid filepath - I tested by simply pasting the filepath in a
browser and was able to generate the image).

This is going to require some debugging, which will have to wait
until Monday because two days of trying to geth the wkhtmltopdf
C library to bind correctly has really taken a lot out of me.